Move a bunch of methods to LibraryHelper module
